+++ src/libaktualizr-posix/ipuptanesecondary.h
@@ -10,11 +10,9 @@ namespace Uptane {
 
 class IpUptaneSecondary : public SecondaryInterface {
  public:
-  static std::pair<bool, std::shared_ptr<Uptane::SecondaryInterface>> connectAndCreate(const std::string& address,
-                                                                                       unsigned short port);
+  static SecondaryInterface::Ptr connectAndCreate(const std::string& address, unsigned short port);
 
-  static std::pair<bool, std::shared_ptr<Uptane::SecondaryInterface>> create(const std::string& address,
-                                                                             unsigned short port, int con_fd);
+  static SecondaryInterface::Ptr create(const std::string& address, unsigned short port, int con_fd);
 
   explicit IpUptaneSecondary(const std::string& address, unsigned short port, EcuSerial serial,
                              HardwareIdentifier hw_id, PublicKey pub_key);
@@ -28,7 +26,8 @@ class IpUptaneSecondary : public SecondaryInterface {
   int32_t getRootVersion(bool /* director */) const override { return 0; }
   bool putRoot(const std::string& /* root */, bool /* director */) override { return true; }
   bool sendFirmware(const std::string& data) override;
-  Json::Value getManifest() const override;
+  data::ResultCode::Numeric install(const std::string& target_name) override;
+  Manifest getManifest() const override;
 
  private:
   const std::pair<std::string, uint16_t>& getAddr() const { return addr_; }