+++ CHANGELOG.md
@@ -6,6 +6,11 @@ Our versioning scheme is `YEAR.N` where `N` is incremented whenever a new releas
 
 ## [??? (unreleased)]
 
+### Added
+
+- Ostree update on IP Secondaries: [PR](https://github.com/advancedtelematic/aktualizr/pull/1518)
+- Basic file update on IP Secondaries: [PR](https://github.com/advancedtelematic/aktualizr/pull/1518)
+
 ## [2019.11] - 2019-12-12
 
 ### Added